add home manager
add

add

fix

many more pkgs, an attempt at home manager with little success...

fix home manager path issue

fix depr warning

delete starship

+++ modules/zsh.nix
@@ -8,11 +8,26 @@
 
 {
   programs.zsh = {
+    enable = true;
+    enableCompletion = true;
+    # initExtra = ''
+    #   export PATH="$PATH:$HOME/bin:$HOME/.local/bin:$HOME/go/bin"
+    # '';
+    # oh-my-zsh = {
+    #   enable = true;
+    #   theme = "robbyrussell";
+    #   plugins = [
+    #     "sudo"
+    #     "terraform"
+    #     "systemadmin"
+    #     "vi-mode"
+    #   ];
+    # };
     # ohMyZsh = {
     #   enable = true;
     #   plugins = [ "git" "sudo" "docker" "kubectl" ];
     # };
-    enable = true;
+#     enable = true;
     interactiveShellInit = ''
 # First, ensure completion system is initialized
 autoload -Uz compinit
@@ -65,6 +80,7 @@ function aug-env {
 
 
 ## Aliases
+alias l="ls -la"
 alias ll="ls -la"
 alias lg="lazygit"
 alias ld="lazydocker"
@@ -86,13 +102,63 @@ alias yessleep="sudo pmset -b disablesleep 0"
 alias nre="darwin-rebuild switch --flake ~/nix#m4 --show-trace -v"
 
 ## python
-function poetry-venv-re {
-  rm -rf venv
-  python3.11 -m venv "venv"
+function poetry-venv {
+  TARGET_DIR=$1
+  if [ "$#" -ne 1 ]; then
+    TARGET_DIR="."
+  fi
+
+  ORIGINAL_DIR=$(pwd)
+
+  if [ ! -d "$TARGET_DIR" ]; then
+    echo "Error: Directory '$TARGET_DIR' does not exist"
+    return 1
+  fi
+
+  if [ ! -d "$TARGET_DIR/venv" ]; then
+    echo "Error: No venv found in '$TARGET_DIR'"
+    return 1
+  fi
+
+  cd "$TARGET_DIR"
   source ./venv/bin/activate
   poetry env info
-  poetry install
+  cd "$ORIGINAL_DIR"
 }
-    '';
+
+function poetry-venv-re {
+  if [ "$#" -lt 1 ] || [ "$#" -gt 2 ]; then
+    echo "Usage: poetry-venv-re PYTHON_VERSION [--no-root]"
+    echo "Example: poetry-venv-re python3.11 --no-root"
+    return 1
+  fi
+
+  PYTHON_VERSION=$1
+  
+  # Validate python version arg
+  if ! command -v "$PYTHON_VERSION" >/dev/null 2>&1; then
+    echo "Error: Python version '$PYTHON_VERSION' not found"
+    return 1
+  fi
+
+  # Validate --no-root flag if provided
+  if [ "$#" -eq 2 ] && [ "$2" != "--no-root" ]; then
+    echo "Error: Second argument must be '--no-root' if provided"
+    return 1
+  fi
+
+  rm -rf venv
+  $PYTHON_VERSION -m venv "venv"
+  source ./venv/bin/activate
+  poetry env info
+  poetry install $2  # Will be either empty or "--no-root"
+}
+
+## docker
+# when installed with advanced/instsall-in-home-dir
+# putting it first so it doesnt conflict
+# https://docs.docker.com/desktop/setup/install/mac-install/
+export PATH=$HOME/.docker/bin:$PATH
+  '';
   };
 }
